add undo compressor log support
