--- pkg/compressor/compressor_factory.go
@@ -1,39 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package compressor
-
-func (c CompressorType) GetCompressor() Compressor {
-	switch c.String() {
-	case CompressorNone.String():
-		return &NoneCompressor{}
-	case CompressorGzip.String():
-		return &Gzip{}
-	case CompressorZip.String():
-		return &Zip{}
-	case CompressorBzip2.String():
-		return &Bzip2{}
-	case CompressorLz4.String():
-		return &Lz4{}
-	case CompressorZstd.String():
-		return &Zstd{}
-	case CompressorDeflate.String():
-		return &DeflateCompress{}
-	default:
-		panic("compressor type not implement")
-	}
-}
