--- pkg/compressor/compressor_type_test.go
@@ -1,35 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package compressor
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestCompressorType(t *testing.T) {
-	assert.Equal(t, CompressorNone, CompressorType(0))
-	assert.Equal(t, CompressorGzip, CompressorType(1))
-	assert.Equal(t, CompressorZip, CompressorType(2))
-	assert.Equal(t, CompressorSevenz, CompressorType(3))
-	assert.Equal(t, CompressorBzip2, CompressorType(4))
-	assert.Equal(t, CompressorLz4, CompressorType(5))
-	assert.Equal(t, CompressorDeflate, CompressorType(6))
-	assert.Equal(t, CompressorZstd, CompressorType(7))
-}
